Renaming file events.js
